Show ranged/melee toggle for axes
